# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hub William API",
    "description": "Control-plane and gateway API for Hub William agent pools",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "Access and refresh sessions rotated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh session missing or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthenticatedUser": {
        "type": "object",
        "required": ["id", "username"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "recovery_email": {
            "type": ["string", "null"]
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "SessionResponse": {
        "type": "object",
        "required": ["user"],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUser"
          }
        }
      },
...
    }
  },
...
}